add session create and load shortcuts
Add <leader>sn and <leader>sl, for 'session new' and 'session load'.
These shortcuts create a session based on the current directory, which
can be loaded from even a new vim session in the same directory at a later
time. The session files are stored in ~/.vim/sessions, which must exist
for these shortcuts to work.